+++ tkk_game/src/main/resources/templates/waiting.html
@@ -11,7 +11,7 @@
       console.log('待機中... Player Name:', playerName);
       
       // SSE接続を開始
-      const sse = new EventSource('/waitRoom');
+      const sse = new EventSource('/match/waitRoom');
       sse.onmessage = function (event) {
         const rooms = JSON.parse(event.data);
         console.log('待機室:', rooms);
@@ -52,10 +52,10 @@
   <div id="haveRequest" style="display: none;">
     <h2>対戦リクエストが届いています！</h2>
     <p><span id="P2Name"></span> さんから対戦リクエストがあります。</p>
-    <form th:action="@{/acceptMatch}" method="post" style="display: inline;">
+    <form th:action="@{/match/accept}" method="post" style="display: inline;">
       <button type="submit">承認する</button>
     </form>
-    <form th:action="@{/rejectMatch}" method="post" style="display: inline;">
+    <form th:action="@{/match/reject}" method="post" style="display: inline;">
       <button type="submit">拒否する</button>
     </form>
   </div>
